Moved claims to theme module
