+++ core/store/modules/index.ts
@@ -1,19 +1,15 @@
-// import category from './category'
 import payment from './payment'
 import shipping from './shipping'
 import ui from './ui-store'
 import checkout from './checkout'
-import claims from './claims'
 import sync from './sync'
 import promoted from './promoted-offers'
 
 export default {
-  // category,
   payment,
   shipping,
   ui,
   checkout,
-  claims,
   sync,
   promoted
 }
