+++ src/modules/claims/index.ts
@@ -1,0 +1,9 @@
+import { VueStorefrontModule } from '@vue-storefront/module'
+import { module } from './store'
+
+const KEY = 'claims'
+
+export const Claims = new VueStorefrontModule({
+  key: KEY,
+  store: { modules: [{ key: KEY, module }]}
+})
