+++ src/modules/index.ts
@@ -10,6 +10,7 @@ import { Example } from '@vue-storefront/core/modules/module-template'
 import { Notification } from '@vue-storefront/core/modules/notification'
 import { RecentlyViewed } from '@vue-storefront/core/modules/recently-viewed'
 import { Homepage } from "./homepage"
+import { Claims } from './claims'
 // Some modules  that still needs API refactoring are  temporary registered in core
 // This is how you can adjust any module with application-specific behavior
 const extendedExample = new VueStorefrontModule({
